+++ src/js/lang/default.js
@@ -1,4 +1,6 @@
 export default {
+    availableFilterHeader: 'Filter available',
+    availableHeader: 'Available',
     moveLeft: 'Move left',
     moveAllLeft: 'Move all left',
     moveRight: 'Move right',
@@ -7,4 +9,6 @@ export default {
     moveUp: 'Move item up',
     noAvailableOptions: 'No available options',
     noSelectedOptions: 'No selected options',
+    selectedFilterHeader: 'Filter selected',
+    selectedHeader: 'Selected',
 };