+++ src/js/shapes/languageShape.js
@@ -6,8 +6,12 @@ export default PropTypes.shape({
     moveRight: PropTypes.string.isRequired,
     moveAllRight: PropTypes.string.isRequired,
     // Optional, so not required
+    availableFilterHeader: PropTypes.string,
+    availableHeader: PropTypes.string,
     moveDown: PropTypes.string,
     moveUp: PropTypes.string,
     noAvailableOptions: PropTypes.string,
     noSelectedOptions: PropTypes.string,
+    selectedFilterHeader: PropTypes.string,
+    selectedHeader: PropTypes.string,
 });